Move RegionCoder out to separate trait
