+++ geospace-http/src/main/scala/com/socrata/geospace/http/GeospaceServlet.scala
@@ -9,25 +9,20 @@ import com.socrata.geospace.http.config.GeospaceConfig
 import com.socrata.geospace.http.curatedregions.{CuratedRegionIndexer, CuratedRegionSuggester}
 import com.socrata.geospace.http.ingestion.FeatureIngester
 import com.socrata.geospace.lib.feature.FeatureValidator
-import com.socrata.geospace.lib.regioncache.{SpatialRegionCache, HashMapRegionCache}
 import com.socrata.geospace.lib.shapefile.{SingleLayerShapefileReader, ZipFromArray}
-import com.socrata.geospace.lib.regioncache.RegionCacheKey
 import com.socrata.soda.external.SodaFountainClient
 import com.socrata.soql.types.SoQLMultiPolygon
 import com.socrata.thirdparty.metrics.Metrics
 import javax.servlet.http.{HttpServletResponse => HttpStatus}
-import org.geoscript.geometry.builder
 import org.scalatra._
 import org.scalatra.servlet.FileUploadSupport
 import scala.collection.JavaConverters._
-import scala.concurrent.Future
 
-class GeospaceServlet(sodaFountain: SodaFountainClient,
+class GeospaceServlet(val sodaFountain: SodaFountainClient,
                       coreServer: CoreServerClient,
                       myConfig: GeospaceConfig) extends GeospaceMicroserviceStack
-                      with FileUploadSupport with Metrics with GeospaceParams {
-  val spatialCache = new SpatialRegionCache(myConfig.cache)
-  val stringCache  = new HashMapRegionCache(myConfig.cache)
+                      with FileUploadSupport with Metrics with GeospaceParams with RegionCoder {
+  val cacheConfig = myConfig.cache
 
   // Metrics
   val geocodingTimer = metrics.timer("geocoding-requests")
@@ -127,16 +122,6 @@ class GeospaceServlet(sodaFountain: SodaFountainClient,
     }
   }
 
-  // Given points, encode them with SpatialIndex and return a sequence of IDs, "" if no matching region
-  // Also describe how the getting the region file is async and thus the coding happens afterwards
-  private def geoRegionCode(resourceName: String, points: Seq[Seq[Double]]): Future[Seq[Option[Int]]] = {
-    val geoPoints = points.map { case Seq(x, y) => builder.Point(x, y) }
-    val futureIndex = spatialCache.getFromSoda(sodaFountain, resourceName)
-    futureIndex.map { index =>
-      geoPoints.map { pt => index.firstContains(pt).map(_.item) }
-    }
-  }
-
   post("/v1/regions/:resourceName/stringcode") {
     val strings = parsedBody.extract[Seq[String]]
     if (strings.isEmpty) halt(HttpStatus.SC_BAD_REQUEST,
@@ -146,11 +131,6 @@ class GeospaceServlet(sodaFountain: SodaFountainClient,
     new AsyncResult { val is =
       geocodingTimer.time { stringCode(resourceName, column, strings) }
     }
-  }
-
-  private def stringCode(resourceName: String, columnName: String, strings: Seq[String]): Future[Seq[Option[Int]]] = {
-    val futureIndex = stringCache.getFromSoda(sodaFountain, RegionCacheKey(resourceName, columnName))
-    futureIndex.map { index => strings.map { str => index.get(str.toLowerCase) } }
   }
 
   // scalastyle:off
